requirements.txt
requests
pandas
networkx
python-dotenv
matplotlib
seaborn
GitPython
seaborn
tqdm
scikit-learn
google-generativeai
google-cloud-aiplatform
vertexai